requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
torch
ultralytics
huggingface_hub
opencv-python-headless
Pillow
numpy
